+++ R/allGenerics.R
@@ -32,7 +32,7 @@
 #'     dimnames =
 #'         list(paste0("sampleX", seq_len(n)), paste0("Y", seq_len(k)))
 #' )
-#' resModtTestJoint <- sbivar(X, Y2, Cx, method = "Modified")
+#' resModtTestJoint <- sbivar(X, Y2, Cx, method = "Modified t-test")
 #' # Single image analysis on synthetic data, converted to SpatialExperiment
 #' if (require(SpatialExperiment)) {
 #'     seX <- SpatialExperiment(